Add database
We keep separate tables for each setting / preference so as to
simplify the logic for applying the stored room setting or falling
back to the configured default. Otherwise we may have to additionally
check for NULL in scenarios where only one of the settings matter.

May have to revisit latter, especially if the performance penalty of
potentially hitting the database twice in some scenarios is a concern.